cors policy corrected!

+++ src/routes/authorRouter.ts
@@ -5,9 +5,14 @@ import logger from '../other_services/winstonLogger';
 import jwt, { JwtPayload } from "jsonwebtoken";
 
 const router = express.Router();
+
+const myCors = (req : any, res : any, next: any) =>{
+    res.setHeader("Access-Control-Allow-Origin", "*");
+    next();
+}
  
 // -------------------- Get author by id --------------------
-router.get("/author/:id", async (req: Request<{ id: number}>, res) => {
+router.get("/author/:id", myCors, async (req: Request<{ id: number}>, res) => {
     try {
         const result: Author = await getAuthorById(req.params.id);
         res.status(200).send(result);
@@ -17,7 +22,7 @@ router.get("/author/:id", async (req: Request<{ id: number}>, res) => {
     }
 });
 
-//FIXME: why are we doing | null here?
+
 export async function getAuthorById(id: number) {
     try{
         const result: Author | null = await Author.findByPk(id);
@@ -29,9 +34,10 @@ export async function getAuthorById(id: number) {
     }
 }
 
+
 // create author
 // TODO: look into making total_books set automatically prop with trigegrs
-router.post("/author",  async (req, res) => {
+router.post("/author", myCors, async (req, res) => {
     try{
         const result: Author = await createAuthor(req.body);
         res.status(200).json(result);
@@ -56,7 +62,7 @@ export async function createAuthor(values: Author) {
 
 
 // get all books for author
-router.get("/author/:id/books", async (req: Request<{ id: number}>, res) => {
+router.get("/author/:id/books", myCors, async (req: Request<{ id: number}>, res) => {
     try {
         const result: Book[] = await getAuthorBooks(req.params.id);
         res.status(200).send(result);
@@ -82,8 +88,9 @@ export async function getAuthorBooks(id: number) {
     }
 }
 
+
 // Get favorited author
-router.get("/user/:id/author/:authorid", async (req, res) => {
+router.get("/user/:id/author/:authorid", myCors, async (req, res) => {
     try {
         const result = await getFavoritedAuthor(req.params.id, req.params.authorid);
         res.status(200).send(result);
@@ -108,8 +115,9 @@ export async function getFavoritedAuthor(id: string, authorid: string) {
     }
 }
 
+
 // Create favorited author
-router.post("/user/:userid/author/:authorid", async (req, res) => {
+router.post("/user/:userid/author/:authorid", myCors, async (req, res) => {
     try {
         const clones = await FavoritedAuthor.findAll({
             where: {
@@ -131,6 +139,7 @@ router.post("/user/:userid/author/:authorid", async (req, res) => {
 
 export async function createFavoritedAuthor(id: string, authorid: string) {
     try{
+   
         const result: FavoritedAuthor = await FavoritedAuthor.create({
             author_id: authorid,
             user_id: id,
@@ -143,10 +152,22 @@ export async function createFavoritedAuthor(id: string, authorid: string) {
 }
 
 
-//Error with deleting a favorited connection Error: ER_NO_REFERENCED_ROW_2: Cannot delete or update a parent row: a foreign key constraint fails (`bookstore`.`favorited_authors`, CONSTRAINT `favorited_authors_ibfk_2` FOREIGN KEY (`author_id`) REFERENCES `authors` (`id`))
-//TODO: Fix this error
+
+
 router.delete("/user/:userid/author/:authorid", async (req, res) => {
     try {
+        const authors = await FavoritedAuthor.findAll({
+            where: {
+                user_id: req.params.userid,
+                author_id: req.params.authorid,
+            }
+        });
+
+        if(authors.length === 0) {
+            res.status(200).send("No favorited author to delete");
+            return;
+        }
+
         const result = await deleteFavoritedAuthor(req.params.userid, req.params.authorid);
         res.status(200).send(result);
     } catch (error) {
@@ -156,18 +177,36 @@ router.delete("/user/:userid/author/:authorid", async (req, res) => {
 });
 
 export async function deleteFavoritedAuthor(id: string, authorid: string) {
-    try{
+    try {
+        const array = await FavoritedAuthor.findAll({
+            where: {
+                user_id: id,
+                author_id: authorid,
+            }
+        });
+
+        console.log(array.length)
+        if(array.length === 0) {
+            return "No favorited author to delete";
+        }
         const result: number = await FavoritedAuthor.destroy({
+          
             where: {
                 author_id: authorid,
                 user_id: id,
             }
         });
+        
+        
         return result;
-    } catch (error){
+    } catch (error) {
+        console.log(error);
         logger.error("Error with deleting a favorited connection", error);
         throw error;
     }
 }
+
+
+
 
 export default router;
